Fix a csv constructor bug
The constructor only assigns the colunms to the hashtable if the lgo file dows not exists. Correct it to verify if the hashtable has the columns and assign to it a empty value. Now it´s possible to assign value to the column direct through the hashtable and column name.